fix feeds tss encoding
